Update to tests to fixe the few tests that were failing
It was mostly instanciation issues due to the change from nosetest to pytest. It is cleaned now. The only strange thing left is that test_na_iir now requires an error_threshold of 0.2 and not 0.03. I thought at one point that the issue was due to avg_per_point which had become average_per_point but that does  not seem to be the problem the max_error is still around 0.17. This would probably require more investigations. Probably a parameter is not set the same way as before, there are some saturations warning coming up.