fixing tests
Fixed old issues due to the switch from nosetest to pytest. Now all tests in a session run on the same pyrpl_instance. It's much faster and easier for remotes tests perfomed by github actions runner. We avoid creating too many new connections to the rp. The only issue left is that some tests mess upo the config file for test_lockbox and test_na so they need to be rerun afterwards for now.